Remove self._frames from all classes
