+++ tools/manual/frameviewer/frame.py
@@ -7,14 +7,11 @@ from tkinter import ttk, TclError
 from functools import partial
 from time import time
 
-import cv2
-import numpy as np
-from PIL import Image, ImageTk
-
 from lib.gui.control_helper import set_slider_rounding
 from lib.gui.custom_widgets import Tooltip
 from lib.gui.utils import get_images
 
+from .control import Navigation, BackgroundImage
 from .editor import (BoundingBox, ExtractBox, Landmarks, Mask, # noqa pylint:disable=unused-import
                      Mesh, View)
 
@@ -30,20 +27,16 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         The paned window that the display frame resides in
     tk_globals: :class:`~tools.manual.manual.TkGlobals`
         The tkinter variables that apply to the whole of the GUI
-    frames: :class:`FrameNavigation`
-        The object that holds the cache of frames.
     detected_faces: :class:`tools.manual.detected_faces.DetectedFaces`
         The detected faces stored in the alignments file
     """
-    def __init__(self, parent, tk_globals, frames, detected_faces):
-        logger.debug("Initializing %s: (parent: %s, tk_globals: %s, frames: %s, "
-                     "detected_faces: %s)",
-                     self.__class__.__name__, parent, tk_globals, frames, detected_faces)
+    def __init__(self, parent, tk_globals, detected_faces):
+        logger.debug("Initializing %s: (parent: %s, tk_globals: %s, detected_faces: %s)",
+                     self.__class__.__name__, parent, tk_globals, detected_faces)
         super().__init__(parent)
         self.pack(side=tk.LEFT, anchor=tk.NW)
 
         self._globals = tk_globals
-        self._frames = frames
         self._det_faces = detected_faces
 
         self._actions_frame = ActionsFrame(self)
@@ -59,7 +52,6 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         self._canvas = FrameViewer(self._video_frame,
                                    self._globals,
                                    self._det_faces,
-                                   self._frames,
                                    self._actions_frame.actions,
                                    self._actions_frame.tk_selected_action)
 
@@ -67,6 +59,7 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         self._transport_frame.pack(side=tk.BOTTOM, padx=5, pady=5, fill=tk.X)
 
         self._nav = self._add_nav()
+        self._navigation = Navigation(self)
         self._buttons = self._add_transport()
         self._add_transport_tk_trace()
         self._actions_frame.add_optional_buttons(self.editors)
@@ -87,11 +80,11 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
     @property
     def _btn_action(self):
         """ dict: {`name`: `action`} Command lookup for navigation buttons """
-        actions = dict(play=self.handle_play_button,
-                       beginning=self.goto_first_frame,
-                       prev=self.decrement_frame,
-                       next=self.increment_frame,
-                       end=self.goto_last_frame,
+        actions = dict(play=self._navigation.handle_play_button,
+                       beginning=self._navigation.goto_first_frame,
+                       prev=self._navigation.decrement_frame,
+                       next=self._navigation.increment_frame,
+                       end=self._navigation.goto_last_frame,
                        save=self._det_faces.save)
         return actions
 
@@ -111,6 +104,12 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         return self._canvas.editors
 
     @property
+    def navigation(self):
+        """ :class:`~tools.manual.frameviewer.control.Navigation`: Class that handles frame
+        Navigation and transport. """
+        return self._navigation
+
+    @property
     def tk_control_colors(self):
         """ :dict: Editor key with :class:`tkinter.StringVar` containing the selected color hex
         code for each annotation """
@@ -128,24 +127,14 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         return self._canvas.colors
 
     @property
-    def _frames_count(self):
-        """ int: The number of frames based on the current navigation mode """
-        return self._det_faces.filter.count
-
-    @property
     def _filter_modes(self):
         """ list: The filter modes combo box values """
         return ["All Frames", "Has Face(s)", "No Faces", "Multiple Faces"]
 
-    @property
-    def _current_nav_frame_count(self):
-        """ int: The current frame count for the transport slider """
-        return self._nav["scale"].cget("to") + 1
-
     def _add_nav(self):
         """ Add the slider to navigate through frames """
         self._globals.tk_transport_index.trace("w", self._set_frame_index)
-        max_frame = self._frames.frame_count - 1
+        max_frame = self._globals.frame_count - 1
 
         frame = ttk.Frame(self._transport_frame)
 
@@ -212,13 +201,13 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
 
     def _add_transport_tk_trace(self):
         """ Add the tkinter variable traces to buttons """
-        self._frames.tk_is_playing.trace("w", self._play)
+        self._navigation.tk_is_playing.trace("w", self._play)
         self._det_faces.tk_unsaved.trace("w", self._toggle_save_state)
 
     def _add_filter_mode_combo(self, frame):
         """ Add the navigation mode combo box to the transport frame """
         self._globals.tk_filter_mode.set("All Frames")
-        self._globals.tk_filter_mode.trace("w", self._nav_scale_callback)
+        self._globals.tk_filter_mode.trace("w", self._navigation.nav_scale_callback)
         nav_frame = ttk.Frame(frame)
         lbl = ttk.Label(nav_frame, text="Filter:")
         lbl.pack(side=tk.LEFT, padx=(0, 5))
@@ -236,38 +225,6 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         idx = (self._filter_modes.index(current_mode) + 1) % len(self._filter_modes)
         self._globals.tk_filter_mode.set(self._filter_modes[idx])
 
-    def _nav_scale_callback(self, *args, reset_progress=True):  # pylint:disable=unused-argument
-        """ Adjust transport slider scale for different filters.
-
-        Returns
-        -------
-        bool
-            ``True`` if the navigation scale has been updated otherwise ``False``
-        """
-        if reset_progress:
-            self._frames.stop_playback()
-        if self._current_nav_frame_count == self._frames_count:
-            logger.trace("Filtered count has not changed. Returning")
-            return False
-        max_frame = max(0, self._frames_count - 1)
-        logger.debug("Filtered frame count has changed. Updating from %s to %s",
-                     self._current_nav_frame_count, self._frames_count)
-        self._nav["scale"].config(to=max_frame)
-        self._nav["label"].config(text="/{}".format(max_frame))
-        state = "disabled" if max_frame == 0 else "normal"
-        self._nav["entry"].config(state=state)
-        if reset_progress:
-            self._globals.tk_transport_index.set(0)
-        return True
-
-    def handle_play_button(self):
-        """ Handle the play button.
-
-        Switches the :attr:`_frames.is_playing` variable.
-        """
-        is_playing = self._frames.tk_is_playing.get()
-        self._frames.tk_is_playing.set(not is_playing)
-
     def set_action(self, key):
         """ Set the current action based on keyboard shortcut
 
@@ -284,7 +241,7 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
     def _play(self, *args, frame_count=None):  # pylint:disable=unused-argument
         """ Play the video file. """
         start = time()
-        is_playing = self._frames.tk_is_playing.get()
+        is_playing = self._navigation.tk_is_playing.get()
         icon = "pause" if is_playing else "play"
         self._buttons["play"].config(image=get_images().icons[icon])
 
@@ -293,8 +250,8 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
             return
 
         # Populate the filtered frames count on first frame
-        frame_count = self._frames_count if frame_count is None else frame_count
-        self.increment_frame(frame_count=frame_count, is_playing=True)
+        frame_count = self._det_faces.filter.count if frame_count is None else frame_count
+        self._navigation.increment_frame(frame_count=frame_count, is_playing=True)
         delay = 16  # Cap speed at approx 60fps max. Unlikely to hit, but just in case
         duration = int((time() - start) * 1000)
         delay = max(1, delay - duration)
@@ -306,75 +263,6 @@ class DisplayFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
         """ Toggle the state of the save button when alignments are updated. """
         state = ["!disabled"] if self._det_faces.tk_unsaved.get() else ["disabled"]
         self._buttons["save"].state(state)
-
-    def increment_frame(self, frame_count=None, is_playing=False):
-        """ Update The frame navigation position to the next frame based on filter. """
-        if not is_playing:
-            self._frames.stop_playback()
-        position = self._globals.tk_transport_index.get()
-        face_count_change = self._check_face_count_change()
-        if face_count_change:
-            position -= 1
-        frame_count = self._frames_count if frame_count is None else frame_count
-        if not face_count_change and (frame_count == 0 or position == frame_count - 1):
-            logger.debug("End of Stream. Not incrementing")
-            self._frames.stop_playback()
-            return
-        self._globals.tk_transport_index.set(min(position + 1, max(0, frame_count - 1)))
-
-    def decrement_frame(self):
-        """ Update The frame navigation position to the previous frame based on filter. """
-        self._frames.stop_playback()
-        position = self._globals.tk_transport_index.get()
-        face_count_change = self._check_face_count_change()
-        if face_count_change:
-            position += 1
-        if not face_count_change and (self._frames_count == 0 or position == 0):
-            logger.debug("End of Stream. Not incrementing")
-            return
-        self._globals.tk_transport_index.set(min(max(0, self._frames_count - 1),
-                                                 max(0, position - 1)))
-
-    def _check_face_count_change(self):
-        """ Check whether the face count for the current filter has changed, and update the
-        transport scale appropriately.
-
-        Perform additional check on whether the current frame still meets the selected navigation
-        mode filter criteria.
-
-        Returns
-        -------
-        bool
-            ``True`` if the currently active frame no longer meets the filter criteria otherwise
-            ``False``
-        """
-        filter_mode = self._globals.filter_mode
-        if filter_mode not in ("No Faces", "Multiple Faces"):
-            return False
-        if not self._nav_scale_callback(reset_progress=False):
-            return False
-        face_count = len(self._det_faces.current_faces[self._globals.frame_index])
-        if (filter_mode == "No Faces" and face_count != 0) or (filter_mode == "Multiple Faces"
-                                                               and face_count < 2):
-            return True
-        return False
-
-    def goto_first_frame(self):
-        """ Go to the first frame that meets the filter criteria. """
-        self._frames.stop_playback()
-        position = self._globals.tk_transport_index.get()
-        if position == 0:
-            return
-        self._globals.tk_transport_index.set(0)
-
-    def goto_last_frame(self):
-        """ Go to the last frame that meets the filter criteria. """
-        self._frames.stop_playback()
-        position = self._globals.tk_transport_index.get()
-        frame_count = self._frames_count
-        if position == frame_count - 1:
-            return
-        self._globals.tk_transport_index.set(frame_count - 1)
 
 
 class ActionsFrame(ttk.Frame):  # pylint:disable=too-many-ancestors
@@ -663,22 +551,19 @@ class FrameViewer(tk.Canvas):  # pylint:disable=too-many-ancestors
         The tkinter variables that apply to the whole of the GUI
     detected_faces: :class:`AlignmentsData`
         The alignments data for this manual session
-    frames: :class:`FrameNavigation`
-        The frames navigator for this manual session
     actions: tuple
         The available actions from :attr:`ActionFrame.actions`
     tk_action_var: :class:`tkinter.StringVar`
         The variable holding the currently selected action
     """
-    def __init__(self, parent, tk_globals, detected_faces, frames, actions, tk_action_var):
+    def __init__(self, parent, tk_globals, detected_faces, actions, tk_action_var):
         logger.debug("Initializing %s: (parent: %s, tk_globals: %s, detected_faces: %s, "
-                     "frames: %s, actions: %s, tk_action_var: %s)", self.__class__.__name__,
-                     parent, tk_globals, detected_faces, frames, actions, tk_action_var)
+                     "actions: %s, tk_action_var: %s)", self.__class__.__name__,
+                     parent, tk_globals, detected_faces, actions, tk_action_var)
         super().__init__(parent, bd=0, highlightthickness=0, background="black")
         self.pack(side=tk.TOP, fill=tk.BOTH, expand=True, anchor=tk.E)
         self._globals = tk_globals
         self._det_faces = detected_faces
-        self._frames = frames
         self._actions = actions
         self._tk_action_var = tk_action_var
         self._image = BackgroundImage(self)
@@ -857,124 +742,3 @@ class FrameViewer(tk.Canvas):  # pylint:disable=too-many-ancestors
             logger.debug("Binding key '%s' to method %s", key, method)
             self.winfo_toplevel().bind(key, method)
             self.key_bindings[key]["bound_to"] = self.selected_action
-
-
-class BackgroundImage():
-    """ The background image of the canvas """
-    def __init__(self, canvas):
-        self._canvas = canvas
-        self._globals = canvas._globals
-        self._frames = canvas._frames
-        self._det_faces = canvas._det_faces
-        zoom_size = (min(self._globals.frame_display_dims), min(self._globals.frame_display_dims))
-        self._zoom_padding = self._get_padding(zoom_size)
-        placeholder = np.ones((*reversed(self._globals.frame_display_dims), 3), dtype="uint8")
-        self._tk_frame = ImageTk.PhotoImage(Image.fromarray(placeholder))
-        self._tk_face = ImageTk.PhotoImage(Image.fromarray(placeholder))
-        self._image = self._canvas.create_image(self._globals.frame_display_dims[0] / 2,
-                                                self._globals.frame_display_dims[1] / 2,
-                                                image=self._tk_frame,
-                                                anchor=tk.CENTER,
-                                                tags="main_image")
-
-    @property
-    def _current_view_mode(self):
-        """ str: `frame` if global zoom mode variable is set to ``False`` other wise `face`. """
-        retval = "face" if self._globals.is_zoomed else "frame"
-        logger.trace(retval)
-        return retval
-
-    def _get_padding(self, size):
-        """ Obtain the Left, Top, Right, Bottom padding required to place the square face or frame
-        in to the Photo Image
-
-        Returns
-        -------
-        tuple
-            The (Left, Top, Right, Bottom) padding to apply to the face image in pixels
-        """
-        pad_lt = ((self._globals.frame_display_dims[1] - size[0]) // 2,
-                  (self._globals.frame_display_dims[0] - size[1]) // 2)
-        padding = (pad_lt[0],
-                   self._globals.frame_display_dims[1] - size[0] - pad_lt[0],
-                   pad_lt[1],
-                   self._globals.frame_display_dims[0] - size[1] - pad_lt[1])
-        logger.debug("Frame dimensions: %s, size: %s, padding: %s",
-                     self._globals.frame_display_dims, size, padding)
-        return padding
-
-    def refresh(self, view_mode):
-        """ Update the displayed frame.
-
-        Parameters
-        ----------
-        view_mode: ["frame", "face"]
-            The currently active editor's selected view mode.
-        """
-        self._switch_image(view_mode)
-        getattr(self, "_update_tk_{}".format(self._current_view_mode))()
-        logger.trace("Updating background frame")
-
-    def _switch_image(self, view_mode):
-        """ Switch the image between the full frame image and the zoomed face image.
-
-        Parameters
-        ----------
-        view_mode: ["frame", "face"]
-            The currently active editor's selected view mode.
-        """
-        if view_mode == self._current_view_mode:
-            return
-        logger.trace("Switching background image from '%s' to '%s'",
-                     self._current_view_mode, view_mode)
-        img = getattr(self, "_tk_{}".format(view_mode))
-        self._canvas.itemconfig(self._image, image=img)
-        self._globals.tk_is_zoomed.set(view_mode == "face")
-        self._globals.tk_face_index.set(0)
-
-    def _update_tk_face(self):
-        """ Update the currently zoomed face. """
-        face = self._get_zoomed_face()
-        face = cv2.copyMakeBorder(face, *self._zoom_padding, cv2.BORDER_CONSTANT)
-        logger.trace("final shape: %s", face.shape)
-        self._tk_face.paste(Image.fromarray(face))
-
-    def _get_zoomed_face(self):
-        """ Get the zoomed face or a blank image if no faces are available.
-
-        Returns
-        -------
-        :class:`numpy.ndarray`
-            The face sized to the shortest dimensions of the face viewer
-        """
-        frame_idx = self._globals.frame_index
-        face_idx = self._globals.face_index
-        faces_in_frame = self._det_faces.face_count_per_index[frame_idx]
-        size = min(self._globals.frame_display_dims)
-
-        if face_idx + 1 > faces_in_frame:
-            logger.debug("Resetting face index to 0 for more faces in frame than current index: ("
-                         "faces_in_frame: %s, zoomed_face_index: %s", faces_in_frame, face_idx)
-            self._globals.tk_face_index.set(0)
-
-        if faces_in_frame == 0:
-            face = np.ones((size, size, 3), dtype="uint8")
-        else:
-
-            face = self._det_faces.get_face_at_index(frame_idx,
-                                                     self._globals.face_index,
-                                                     self._globals.current_frame["image"],
-                                                     size)
-        logger.trace("face shape: %s", face.shape)
-        return face[..., 2::-1]
-
-    def _update_tk_frame(self):
-        """ Place the currently held frame into :attr:`_tk_frame`. """
-        img = cv2.resize(self._globals.current_frame["image"],
-                         self._globals.current_frame["display_dims"],
-                         interpolation=self._globals.current_frame["interpolation"])[..., 2::-1]
-        padding = self._get_padding(img.shape[:2])
-        if any(padding):
-            img = cv2.copyMakeBorder(img, *padding, cv2.BORDER_CONSTANT)
-        logger.trace("final shape: %s", img.shape)
-        self._tk_frame.paste(Image.fromarray(img))
